fix: #485: Centralize auth session persistence
