+++ apps/web/hooks/auth/useAuthSession.ts
@@ -1,0 +1,5 @@
+import { authStorage } from "@/lib/auth/authStorage";
+
+export const useAuthSession = () => {
+  return authStorage;
+};
