+++ apps/web/lib/auth/storageKeys.ts
@@ -1,0 +1,5 @@
+export const AUTH_STORAGE_KEYS = {
+  accessToken: "kiibee.accessToken",
+  refreshToken: "kiibee.refreshToken",
+  user: "kiibee.user",
+} as const;
